requirements.txt
Flask==2.0.1
flask-socketio
azure-cognitiveservices-speech
eventlet
Werkzeug==2.0.3
